Start testing the actual return types. Used on Iterate.
Previously I actually tried to guess the return values by testing their behavior,
but that lead to a bunch of bugs sneaking in undetected.

Now I actually test the returned iterator types,
their return value types, and so on.

In this commit it is applied on Iterate,
in the follow up commits I'll apply it to all others as well.

This fixed 2 bugs on Iterate:
    - If iterating over a const collection, 'Iterator<X>::begin()' returned X::const_iterator,
      but 'Iterator<X>::iterator' claimed it was 'X::iterator'.
    - 'Iterator<X>::begin() const' returned 'X::iterator', so it
    actually allowed the underlying data to be modified!